Create password component
