+++ frontend/.eslintcache
@@ -1,1 +1,1 @@
-[{"C:\\Users\\karan\\git\\book-project\\frontend\\src\\index.tsx":"1","C:\\Users\\karan\\git\\book-project\\frontend\\src\\reportWebVitals.ts":"2","C:\\Users\\karan\\git\\book-project\\frontend\\src\\login\\Login.tsx":"3","C:\\Users\\karan\\git\\book-project\\frontend\\src\\App.tsx":"4","C:\\Users\\karan\\git\\book-project\\frontend\\src\\register\\Register.tsx":"5"},{"size":506,"mtime":1608916771827,"results":"6","hashOfConfig":"7"},{"size":425,"mtime":1608916282147,"results":"8","hashOfConfig":"7"},{"size":2281,"mtime":1609264037485,"results":"9","hashOfConfig":"7"},{"size":148,"mtime":1609271418373,"results":"10","hashOfConfig":"7"},{"size":2429,"mtime":1609271241758,"results":"11","hashOfConfig":"7"},{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"xlb0qo",{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\karan\\git\\book-project\\frontend\\src\\index.tsx",[],"C:\\Users\\karan\\git\\book-project\\frontend\\src\\reportWebVitals.ts",[],"C:\\Users\\karan\\git\\book-project\\frontend\\src\\login\\Login.tsx",[],"C:\\Users\\karan\\git\\book-project\\frontend\\src\\App.tsx",[],"C:\\Users\\karan\\git\\book-project\\frontend\\src\\register\\Register.tsx",[]]
+[{"C:\\Users\\karan\\git\\book-project\\frontend\\src\\index.tsx":"1","C:\\Users\\karan\\git\\book-project\\frontend\\src\\reportWebVitals.ts":"2","C:\\Users\\karan\\git\\book-project\\frontend\\src\\login\\Login.tsx":"3","C:\\Users\\karan\\git\\book-project\\frontend\\src\\App.tsx":"4","C:\\Users\\karan\\git\\book-project\\frontend\\src\\register\\Register.tsx":"5","C:\\Users\\karan\\git\\book-project\\frontend\\src\\shared\\form\\Password.tsx":"6"},{"size":506,"mtime":1608916771827,"results":"7","hashOfConfig":"8"},{"size":425,"mtime":1608916282147,"results":"9","hashOfConfig":"8"},{"size":1726,"mtime":1609272529611,"results":"10","hashOfConfig":"8"},{"size":148,"mtime":1609272552700,"results":"11","hashOfConfig":"8"},{"size":1742,"mtime":1609272396730,"results":"12","hashOfConfig":"8"},{"size":1679,"mtime":1609272056973,"results":"13","hashOfConfig":"8"},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"xlb0qo",{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\karan\\git\\book-project\\frontend\\src\\index.tsx",[],"C:\\Users\\karan\\git\\book-project\\frontend\\src\\reportWebVitals.ts",[],"C:\\Users\\karan\\git\\book-project\\frontend\\src\\login\\Login.tsx",[],"C:\\Users\\karan\\git\\book-project\\frontend\\src\\App.tsx",[],"C:\\Users\\karan\\git\\book-project\\frontend\\src\\register\\Register.tsx",[],"C:\\Users\\karan\\git\\book-project\\frontend\\src\\shared\\form\\Password.tsx",[]]
